Fix sanitizer in AudioFlinger threadLoop.
The integer sanitizer was catching an unsigned integer overflow on the
calculation for throttleMs. This gets implicitly cast to a signed
integer anyways, so just perform the cast explicitly.

Bug: 30969751
Test: Compiles with and without integer sanitization.
Test: Device boots.
Merged-In: I4ef7c6f2e871ee56b81e87f3d71be8d85b222934

+++ services/audioflinger/Threads.cpp
@@ -3314,7 +3314,7 @@ bool AudioFlinger::PlaybackThread::threadLoop()
                             lastWriteFinished,previousLastWriteFinished, &deltaNs);
                         const int32_t deltaMs = deltaNs / 1000000;
 
-                        const int32_t throttleMs = mHalfBufferMs - deltaMs;
+                        const int32_t throttleMs = (int32_t)mHalfBufferMs - deltaMs;
                         if ((signed)mHalfBufferMs >= throttleMs && throttleMs > 0) {
                             usleep(throttleMs * 1000);
                             // notify of throttle start on verbose log
